VIX-3076 Remove Pixel Specific Properties and Methods from PreviewBaseShape.cs
VIX-3076 Remove Pixel Specific Properties and Methods from PreviewBaseShape.cs
