+++ Modules/Preview/VixenPreview/Shapes/PreviewNet.cs
@@ -12,7 +12,7 @@ using Vixen.Sys;
 namespace VixenModules.Preview.VixenPreview.Shapes
 {
 	[DataContract]
-	public class PreviewNet : PreviewBaseShape
+	public class PreviewNet : PreviewLightBaseShape
 	{
 		[DataMember] private PreviewPoint _topLeft;
 		[DataMember] private PreviewPoint _topRight;
