+++ Modules/Preview/VixenPreview/VixenPreviewControl.cs
@@ -307,9 +307,9 @@ namespace VixenModules.Preview.VixenPreview
 			get
 			{
 				int count = 0;
-				foreach (DisplayItem displayItem in DisplayItems)
-				{
-					count += displayItem.Shape.Pixels.Count;
+				foreach (DisplayItem displayItem in DisplayItems.Where(item => item.IsLightShape()))
+				{					
+					count += displayItem.LightShape.Pixels.Count;				
 				}
 				return count;
 			}
@@ -587,7 +587,7 @@ namespace VixenModules.Preview.VixenPreview
 							return;
 						}
 
-						// Is there a single dislay item selected?
+						// Is there a single display item selected?
 						if (_selectedDisplayItem != null && !controlPressed)
 						{
 							// Lets see if we've got a drag point.
@@ -696,7 +696,7 @@ namespace VixenModules.Preview.VixenPreview
 							if (ItemName != String.Empty)
 							{
 								newDisplayItem.Shape.Name = ItemName + ItemIndex++;
-								newDisplayItem.Shape.PixelSize = ItemBulbSize;
+								newDisplayItem.LightShape.PixelSize = ItemBulbSize;
 							}
 						}
 						else if (_currentTool == Tools.Ellipse)
@@ -1004,9 +1004,12 @@ namespace VixenModules.Preview.VixenPreview
 
 		private void SetDisplayItemZ(DisplayItem displayItem, int pos)
 		{
-			foreach (PreviewPixel p in displayItem.Shape.Pixels)
+			if (displayItem.IsLightShape())
 			{
-				p.Z = pos;
+				foreach (PreviewPixel p in displayItem.LightShape.Pixels)
+				{
+					p.Z = pos;
+				}
 			}
 		}
 
@@ -1124,7 +1127,7 @@ namespace VixenModules.Preview.VixenPreview
 			}
 		}
 
-		protected override void OnKeyDown(KeyEventArgs e)
+		protected override async void OnKeyDown(KeyEventArgs e)
 		{
 			if (e.KeyCode == Keys.Delete)
 			{
@@ -1239,7 +1242,7 @@ namespace VixenModules.Preview.VixenPreview
 						if (elementsForm.SelectedNode == null && modifyType.Equals("AddNew"))
 						{
 							//Intercept and populate the element tree
-							if (ShowElementCreateTemplateForCurrentTool() && elementsForm.SelectedNode != null)
+							if (await ShowElementCreateTemplateForCurrentTool() && elementsForm.SelectedNode != null)
 							{
 								_selectedDisplayItem.Shape.Reconfigure(elementsForm.SelectedNode);
 							}
@@ -1290,7 +1293,7 @@ namespace VixenModules.Preview.VixenPreview
 			base.OnKeyDown(e);
 		}
 
-		public void VixenPreviewControl_MouseUp(object sender, MouseEventArgs e)
+		public async void VixenPreviewControl_MouseUp(object sender, MouseEventArgs e)
 		{
 			if (_mouseCaptured)
 			{
@@ -1312,7 +1315,7 @@ namespace VixenModules.Preview.VixenPreview
 					if (_elementSelected == null && modifyType.Equals("AddNew"))
 					{
 						//Intercept and populate the element tree
-						if (ShowElementCreateTemplateForCurrentTool() && elementsForm.SelectedNode != null)
+						if (await ShowElementCreateTemplateForCurrentTool() && elementsForm.SelectedNode != null)
 						{
 							_selectedDisplayItem.Shape.Reconfigure(elementsForm.SelectedNode);
 						}
@@ -1363,7 +1366,7 @@ namespace VixenModules.Preview.VixenPreview
 			EndUpdate();
 		}
 
-		private bool ShowElementCreateTemplateForCurrentTool()
+		private async Task<bool> ShowElementCreateTemplateForCurrentTool()
 		{
 
 			IElementTemplate template = null;
@@ -1402,7 +1405,7 @@ namespace VixenModules.Preview.VixenPreview
 
 			if (template != null)
 			{
-				success = elementsForm.SetupTemplate(template);
+				success = await elementsForm.SetupTemplate(template);
 			}
 
 			return success;
@@ -1523,59 +1526,10 @@ namespace VixenModules.Preview.VixenPreview
 		public void AddNodeToPixelMapping(DisplayItem item)
 		{
 			if(NodeToPixel == null) NodeToPixel = new ConcurrentDictionary<ElementNode, List<PreviewPixel>>();
-			foreach (PreviewPixel pixel in item.Shape.Pixels)
+
+			if (item.IsLightShape())
 			{
-				if (pixel.Node != null)
-				{
-					List<PreviewPixel> pixels;
-					if (NodeToPixel.TryGetValue(pixel.Node, out pixels))
-					{
-						if (!pixels.Contains(pixel))
-						{
-							pixels.Add(pixel);
-						}
-					}
-					else
-					{
-						pixels = new List<PreviewPixel>();
-						pixels.Add(pixel);
-						NodeToPixel.TryAdd(pixel.Node, pixels);
-					}
-				}
-			}
-		}
-
-		public void RemoveNodeToPixelMapping(DisplayItem item)
-		{
-			if (NodeToPixel == null) NodeToPixel = new ConcurrentDictionary<ElementNode, List<PreviewPixel>>();
-			foreach (PreviewPixel pixel in item.Shape.Pixels)
-			{
-				if (pixel.Node != null)
-				{
-					List<PreviewPixel> pixels;
-					if (NodeToPixel.TryGetValue(pixel.Node, out pixels))
-					{
-						pixels.Remove(pixel);
-					}
-				}
-			}
-		}
-
-		public void Reload()
-		{
-			if (NodeToPixel == null) NodeToPixel = new ConcurrentDictionary<ElementNode, List<PreviewPixel>>();
-			NodeToPixel.Clear();
-			
-			if (DisplayItems == null)
-				throw new System.ArgumentException("DisplayItems == null");
-
-			foreach (DisplayItem item in DisplayItems)
-			{
-				if (item.Shape.Pixels == null)
-					throw new System.ArgumentException("item.Shape.Pixels == null");
-
-
-				foreach (PreviewPixel pixel in item.Shape.Pixels)
+				foreach (PreviewPixel pixel in item.LightShape.Pixels)
 				{
 					if (pixel.Node != null)
 					{
@@ -1592,6 +1546,66 @@ namespace VixenModules.Preview.VixenPreview
 							pixels = new List<PreviewPixel>();
 							pixels.Add(pixel);
 							NodeToPixel.TryAdd(pixel.Node, pixels);
+						}
+					}
+				}
+			}
+		}
+
+		public void RemoveNodeToPixelMapping(DisplayItem item)
+		{
+			if (NodeToPixel == null) NodeToPixel = new ConcurrentDictionary<ElementNode, List<PreviewPixel>>();
+
+			if (item.IsLightShape())
+			{
+				foreach (PreviewPixel pixel in item.LightShape.Pixels)
+				{
+					if (pixel.Node != null)
+					{
+						List<PreviewPixel> pixels;
+						if (NodeToPixel.TryGetValue(pixel.Node, out pixels))
+						{
+							pixels.Remove(pixel);
+						}
+					}
+				}
+			}
+		}
+
+		public void Reload()
+		{
+			if (NodeToPixel == null) NodeToPixel = new ConcurrentDictionary<ElementNode, List<PreviewPixel>>();
+			NodeToPixel.Clear();
+			
+			if (DisplayItems == null)
+				throw new System.ArgumentException("DisplayItems == null");
+
+			foreach (DisplayItem item in DisplayItems)
+			{
+				if (item.IsLightShape() &&
+					item.LightShape.Pixels == null)
+					throw new System.ArgumentException("item.Shape.Pixels == null");
+
+				if (item.IsLightShape())
+				{
+					foreach (PreviewPixel pixel in item.LightShape.Pixels)
+					{
+						if (pixel.Node != null)
+						{
+							List<PreviewPixel> pixels;
+							if (NodeToPixel.TryGetValue(pixel.Node, out pixels))
+							{
+								if (!pixels.Contains(pixel))
+								{
+									pixels.Add(pixel);
+								}
+							}
+							else
+							{
+								pixels = new List<PreviewPixel>();
+								pixels.Add(pixel);
+								NodeToPixel.TryAdd(pixel.Node, pixels);
+							}
 						}
 					}
 				}
@@ -1769,8 +1783,11 @@ namespace VixenModules.Preview.VixenPreview
 			var reverseChange = -info.ChangeAmount;
 			foreach (DisplayItem e in changedPreviewItems)
 			{
-				e.Shape.ResizePixelsBy(reverseChange);
-				info.ChangeAmount = reverseChange;
+				if (e.IsLightShape())
+				{
+					e.LightShape.ResizePixelsBy(reverseChange);
+					info.ChangeAmount = reverseChange;
+				}
 			}
 			EndUpdate();
 		}
@@ -1805,6 +1822,11 @@ namespace VixenModules.Preview.VixenPreview
 				shapes.Add(item.Shape);
 			}
 			return shapes;
+		}
+
+		public List<PreviewLightBaseShape> SelectedLightShapes()
+		{
+			return SelectedShapes().OfType<PreviewLightBaseShape>().ToList();
 		}
 
 		public DisplayItem CreateGroup()
@@ -1848,7 +1870,7 @@ namespace VixenModules.Preview.VixenPreview
 				leftPoint.Add(shape.Shape.Left);
 			}
 			newDisplayItem = new DisplayItem();
-			newDisplayItem.Shape = new PreviewCustom(new PreviewPoint(leftPoint.Min(), topPoint.Min()), SelectedShapes());
+			newDisplayItem.Shape = new PreviewCustom(new PreviewPoint(leftPoint.Min(), topPoint.Min()), SelectedLightShapes());
 			AddDisplayItem(newDisplayItem);
 
 			foreach (DisplayItem item in SelectedDisplayItems)
@@ -1886,7 +1908,7 @@ namespace VixenModules.Preview.VixenPreview
 				
 				if (newDisplayItem != null)
 				{
-					foreach (var previewPixel in newDisplayItem.Shape.Pixels)
+					foreach (var previewPixel in newDisplayItem.LightShape.Pixels)
 					{
 						//Remove any node associations from the template that may have been saved in old versions
 						previewPixel.Node = null;
@@ -1900,7 +1922,7 @@ namespace VixenModules.Preview.VixenPreview
 						//try to associate the template to the nodes by suffix for Previewsingle types
 
 						var children = elementsForm.SelectedNode.GetLeafEnumerator().GroupBy(l => l.Name).Select(x => x.FirstOrDefault()).ToDictionary(x => x.Name);
-						foreach (var previewBaseShape in newDisplayItem.Shape.Strings)
+						foreach (var previewBaseShape in newDisplayItem.LightShape.Strings)
 						{
 							if (previewBaseShape is PreviewSingle)
 							{
@@ -2052,29 +2074,32 @@ namespace VixenModules.Preview.VixenPreview
 
 		public void SeparateTemplateItems(DisplayItem displayItem)
 		{
-			foreach (PreviewBaseShape shape in displayItem.Shape._strings)
+			if (displayItem.IsLightShape())
 			{
-				DisplayItem newDisplayItem;
-				newDisplayItem = new DisplayItem();
-				newDisplayItem.Shape = shape;
-				DisplayItems.Add(newDisplayItem);
-			}
-			if (_selectedDisplayItem != null)
-			{
-				DisplayItems.Remove(_selectedDisplayItem);
-				DeSelectSelectedDisplayItem();
-			}
-			else if (SelectedDisplayItems != null && SelectedDisplayItems.Count > 0)
-			{
-				foreach (DisplayItem item in SelectedDisplayItems)
+				foreach (PreviewLightBaseShape shape in (displayItem.LightShape)._strings)
 				{
-					DisplayItems.Remove(item);
+					DisplayItem newDisplayItem;
+					newDisplayItem = new DisplayItem();
+					newDisplayItem.Shape = shape;
+					DisplayItems.Add(newDisplayItem);
+				}
+				if (_selectedDisplayItem != null)
+				{
+					DisplayItems.Remove(_selectedDisplayItem);
 					DeSelectSelectedDisplayItem();
 				}
-				SelectedDisplayItems.Clear();
-				OnSelectionChanged?.Invoke(this, EventArgs.Empty);
+				else if (SelectedDisplayItems != null && SelectedDisplayItems.Count > 0)
+				{
+					foreach (DisplayItem item in SelectedDisplayItems)
+					{
+						DisplayItems.Remove(item);
+						DeSelectSelectedDisplayItem();
+					}
+					SelectedDisplayItems.Clear();
+					OnSelectionChanged?.Invoke(this, EventArgs.Empty);
+				}
+				EndUpdate();
 			}
-			EndUpdate();
 		}
 
 		#endregion
@@ -2083,14 +2108,16 @@ namespace VixenModules.Preview.VixenPreview
 
 		public void IncreaseBulbSize()
 		{
-			if (_selectedDisplayItem != null && !SelectedShapes().Any())
+			if (_selectedDisplayItem != null &&
+				_selectedDisplayItem.IsLightShape() &&
+				 !SelectedLightShapes().Any())
 			{
-				_selectedDisplayItem.Shape.ResizePixelsBy(1);
+				_selectedDisplayItem.LightShape.ResizePixelsBy(1);
 				PixelResize(1, new List<DisplayItem>( new [] {_selectedDisplayItem}));
 			}
 			else
 			{
-				foreach (PreviewBaseShape shape in SelectedShapes())
+				foreach (PreviewLightBaseShape shape in SelectedLightShapes())
 				{
 					shape.ResizePixelsBy(1);
 				}
@@ -2101,14 +2128,16 @@ namespace VixenModules.Preview.VixenPreview
 
 		public void DecreaseBulbSize()
 		{
-			if (_selectedDisplayItem != null && !SelectedShapes().Any())
+			if (_selectedDisplayItem != null && 
+				_selectedDisplayItem.IsLightShape() &&
+				!SelectedShapes().Any())
 			{
-				_selectedDisplayItem.Shape.ResizePixelsBy(-1);
+				_selectedDisplayItem.LightShape.ResizePixelsBy(-1);
 				PixelResize(1, new List<DisplayItem>(new[] { _selectedDisplayItem }));
 			}
 			else
 			{
-				foreach (PreviewBaseShape shape in SelectedShapes())
+				foreach (PreviewLightBaseShape shape in SelectedLightShapes())
 				{
 					shape.ResizePixelsBy(-1);
 				}
@@ -2120,14 +2149,17 @@ namespace VixenModules.Preview.VixenPreview
 
 		public void MatchBulbSize()
 		{
-			foreach (PreviewBaseShape shape in SelectedShapes())
+			if (SelectedShapes()[0] is PreviewLightBaseShape)
 			{
-				if (shape != SelectedShapes()[0])
+				foreach (PreviewLightBaseShape shape in SelectedLightShapes())
 				{
-					shape.MatchPixelSize(SelectedShapes()[0]);
+					if (shape != SelectedShapes()[0])
+					{
+						shape.MatchPixelSize((PreviewLightBaseShape)SelectedShapes()[0]);
+					}
 				}
+				EndUpdate();
 			}
-			EndUpdate();
 		}
 
 		#endregion
@@ -2383,14 +2415,17 @@ namespace VixenModules.Preview.VixenPreview
 		{
 			foreach (DisplayItem displayItem in displayItems)
 			{
-				foreach (var p in displayItem.Shape.Pixels.Where(pi => pi != null && pi.Node != null))
+				if (displayItem.IsLightShape())
 				{
-					if (!p.Node.Properties.Contains(LocationDescriptor._typeId))
-						p.Node.Properties.Add(LocationDescriptor._typeId);
-					var prop = p.Node.Properties.Get(LocationDescriptor._typeId);
-					((LocationData) prop.ModuleData).X = p.IsHighPrecision ? (int)p.Location.X:p.X + Convert.ToInt32(Data.LocationOffset.X);
-					((LocationData) prop.ModuleData).Y = p.IsHighPrecision ? (int)p.Location.Y:p.Y + Convert.ToInt32(Data.LocationOffset.Y);
-					((LocationData) prop.ModuleData).Y = p.Z;
+					foreach (var p in displayItem.LightShape.Pixels.Where(pi => pi != null && pi.Node != null))
+					{
+						if (!p.Node.Properties.Contains(LocationDescriptor._typeId))
+							p.Node.Properties.Add(LocationDescriptor._typeId);
+						var prop = p.Node.Properties.Get(LocationDescriptor._typeId);
+						((LocationData)prop.ModuleData).X = p.IsHighPrecision ? (int)p.Location.X : p.X + Convert.ToInt32(Data.LocationOffset.X);
+						((LocationData)prop.ModuleData).Y = p.IsHighPrecision ? (int)p.Location.Y : p.Y + Convert.ToInt32(Data.LocationOffset.Y);
+						((LocationData)prop.ModuleData).Y = p.Z;
+					}
 				}
 			}
 		}
